+++ packages/galpy/test_galpy.py
@@ -3,20 +3,28 @@ from functools import reduce
 import pytest
 from pytest_pyodide import run_in_pyodide
 
+NODE_XFAIL_REASON = (
+    "galpy loads matplotlib and there are no supported matplotlib backends on node"
+)
+
 # Need to skip_refcount_check because we use matplotlib
-DECORATORS = [
-    pytest.mark.xfail_browsers(
-        node="galpy loads matplotlib and there are no supported matplotlib backends on node"
-    ),
-    pytest.mark.skip_refcount_check,
-]
 
 
-def galpy_test_decorator(f):
-    return reduce(lambda x, g: g(x), DECORATORS, f)
+def galpy_test_decorator(**kwargs):
+    def dec(f):
+        return reduce(
+            lambda x, g: g(x),
+            [
+                pytest.mark.xfail_browsers(node=NODE_XFAIL_REASON, **kwargs),
+                pytest.mark.skip_refcount_check,
+            ],
+            f,
+        )
+
+    return dec
 
 
-@galpy_test_decorator
+@galpy_test_decorator(firefox="times out")
 @run_in_pyodide(
     packages=[
         "galpy",
@@ -36,7 +44,7 @@ def test_integrate(selenium):
     return None
 
 
-@galpy_test_decorator
+@galpy_test_decorator(firefox="times out")
 @run_in_pyodide(
     packages=[
         "galpy",
@@ -62,7 +70,7 @@ def test_actionAngle(selenium):
     return None
 
 
-@galpy_test_decorator
+@galpy_test_decorator(firefox="times out")
 @run_in_pyodide(
     packages=[
         "galpy",